requirements.txt
gitdb==4.0.11
GitPython==3.1.43
lizard==1.17.13
numpy==2.0.2
pandas==2.2.3
PyDriller==2.7
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
smmap==5.0.1
types-pytz==2024.2.0.20241003
tzdata==2024.2
